requirements.txt
pandas
apache-airflow-providers-mysql
pymysql
sqlalchemy